enhancement code by implement services

+++ src/app/service/storage.service.ts
@@ -1,25 +1,34 @@
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 
 @Injectable({
   providedIn: 'root',
 })
 export class StorageService {
-  constructor() {}
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
 
   getItem(key: string) {
-    const item = localStorage.getItem(key);
-    return item ? JSON.parse(item) : null;
+    if (isPlatformBrowser(this.platformId)) {
+      const item = localStorage.getItem(key);
+      return item ? JSON.parse(item) : null;
+    }
   }
 
   setItem(key: string, data: any) {
-    localStorage.setItem(key, JSON.stringify(data));
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem(key, JSON.stringify(data));
+    }
   }
 
   removeItem(key: string) {
-    localStorage.removeItem(key);
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.removeItem(key);
+    }
   }
 
   clear() {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
   }
 }
